+++ ceammc/ext/src/random/random_float.h
@@ -30,6 +30,8 @@ public:
 
     void onBang() override;
     void onInlet(size_t n, const AtomListView& lv) override;
+
+    void m_gen(t_symbol* s, const AtomListView& lv);
 };
 
 void setup_random_float();